Trait embedded_io::blocking::Seek

pub trait Seek: Io {
    fn seek(&mut self, pos: SeekFrom) -> Result<u64, Self::Error>;

    fn rewind(&mut self) -> Result<(), Self::Error> { ... }
    fn stream_position(&mut self) -> Result<u64, Self::Error> { ... }
}

Blocking seek within streams.

Semantics are the same as [std::io::Seek], check its documentation for details.

Required Methods

fn seek(&mut self, pos: SeekFrom) -> Result<u64, Self::Error>

Seek to an offset, in bytes, in a stream.

Provided Methods

fn rewind(&mut self) -> Result<(), Self::Error>

Rewind to the beginning of a stream.

fn stream_position(&mut self) -> Result<u64, Self::Error>

Returns the current seek position from the start of the stream.