Trait embedded_io::blocking::Write

pub trait Write: Io {
    fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error>;
    fn flush(&mut self) -> Result<(), Self::Error>;

    fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error> { ... }
    fn write_fmt(
        &mut self, 
        fmt: Arguments<'_>
    ) -> Result<(), WriteFmtError<Self::Error>> { ... }
}

Blocking writer.

Semantics are the same as [std::io::Write], check its documentation for details.

Required methods

fn write(&mut self, buf: &[u8]) -> Result<usize, Self::Error>

Write a buffer into this writer, returning how many bytes were written.

fn flush(&mut self) -> Result<(), Self::Error>

Flush this output stream, ensuring that all intermediately buffered contents reach their destination.

Provided methods

fn write_all(&mut self, buf: &[u8]) -> Result<(), Self::Error>

Write an entire buffer into this writer.

fn write_fmt(
&mut self,
fmt: Arguments<'_>
) -> Result<(), WriteFmtError<Self::Error>>

Write a formatted string into this writer, returning any error encountered.