Trait embedded_io::blocking::BufRead

pub trait BufRead: Io {
    // Required methods
    fn fill_buf(&mut self) -> Result<&[u8], Self::Error>;
    fn consume(&mut self, amt: usize);
}

Blocking buffered reader.

Semantics are the same as [std::io::BufRead], check its documentation for details.

Required Methods

fn fill_buf(&mut self) -> Result<&[u8], Self::Error>

Return the contents of the internal buffer, filling it with more data from the inner reader if it is empty.

fn consume(&mut self, amt: usize)

Tell this buffer that amt bytes have been consumed from the buffer, so they should no longer be returned in calls to fill_buf.