embedded_io

Trait BufRead

pub trait BufRead: Read {
    // Required methods
    fn fill_buf(&mut self) -> Result<&[u8], Self::Error>;
    fn consume(&mut self, amt: usize);
}

Blocking buffered reader.

This trait is the embedded-io equivalent of [std::io::BufRead].

Required Methods

fn fill_buf(&mut self) -> Result<&[u8], Self::Error>

Return the contents of the internal buffer, filling it with more data from the inner reader if it is empty.

If no bytes are currently available to read, this function blocks until at least one byte is available.

If the reader is at end-of-file (EOF), an empty slice is returned. There is no guarantee that a reader at EOF will always be so in the future, for example a reader can stop being at EOF if another process appends more bytes to the underlying file.

fn consume(&mut self, amt: usize)

Tell this buffer that amt bytes have been consumed from the buffer, so they should no longer be returned in calls to fill_buf.